+++ dataflow/manual/examples/CFGConditionalOpExpr.java
@@ -1,0 +1,6 @@
+class Test {
+  int test(boolean b) {
+    int x = b ? this.hashCode() : 5;
+    return x;
+  }
+}