+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/builder/CFGTranslationPhaseOne.java
@@ -2549,23 +2549,19 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
         ExpressionTree trueExprTree = tree.getTrueExpression();
         Node trueExprNode = scan(trueExprTree, p);
         trueExprNode = conditionalExprPromotion(trueExprNode, exprType);
-
         extendWithAssignmentForConditionalExpr(condExprVarTree, trueExprTree, trueExprNode);
-
         extendWithExtendedNode(new UnconditionalJump(merge));
 
         addLabelForNextNode(falseStart);
         ExpressionTree falseExprTree = tree.getFalseExpression();
         Node falseExprNode = scan(falseExprTree, p);
         falseExprNode = conditionalExprPromotion(falseExprNode, exprType);
-
         extendWithAssignmentForConditionalExpr(condExprVarTree, falseExprTree, falseExprNode);
-
         extendWithExtendedNode(new UnconditionalJump(merge));
 
         addLabelForNextNode(merge);
         Pair<IdentifierTree, LocalVariableNode> treeAndLocalVarNode =
-                extendWithVarUseNode(condExprVarTree);
+                buildVarUseNode(condExprVarTree);
         Node node =
                 new TernaryExpressionNode(
                         tree, condition, trueExprNode, falseExprNode, treeAndLocalVarNode.second);
@@ -2586,31 +2582,33 @@ public class CFGTranslationPhaseOne extends TreeScanner<Node, Void> {
     private void extendWithAssignmentForConditionalExpr(
             VariableTree condExprVarTree, ExpressionTree caseExprTree, Node caseExprNode) {
         Pair<IdentifierTree, LocalVariableNode> treeAndLocalVarNode =
-                extendWithVarUseNode(condExprVarTree);
+                buildVarUseNode(condExprVarTree);
 
         AssignmentTree assign =
                 treeBuilder.buildAssignment(treeAndLocalVarNode.first, caseExprTree);
         handleArtificialTree(assign);
 
+        // Build a "synthetic" assignment node, allowing special handling in transfer functions
         AssignmentNode assignmentNode =
-                new AssignmentNode(assign, treeAndLocalVarNode.second, caseExprNode);
+                new AssignmentNode(assign, treeAndLocalVarNode.second, caseExprNode, true);
         assignmentNode.setInSource(false);
         extendWithNode(assignmentNode);
     }
 
     /**
-     * Extend the CFG with a {@link LocalVariableNode} representing a use of some variable
+     * Build a pair of {@link IdentifierTree} and {@link LocalVariableNode} to represent a use of
+     * some variable
      *
      * @param varTree tree for the variable
      * @return a pair whose first element is the synthetic {@link IdentifierTree} for the use, and
      *     whose second element is the {@link LocalVariableNode} representing the use
      */
-    private Pair<IdentifierTree, LocalVariableNode> extendWithVarUseNode(VariableTree varTree) {
+    private Pair<IdentifierTree, LocalVariableNode> buildVarUseNode(VariableTree varTree) {
         IdentifierTree condExprVarUseTree = treeBuilder.buildVariableUse(varTree);
         handleArtificialTree(condExprVarUseTree);
         LocalVariableNode condExprVarUseNode = new LocalVariableNode(condExprVarUseTree);
         condExprVarUseNode.setInSource(false);
-        extendWithNode(condExprVarUseNode);
+        // Do not actually add the node to the CFG
         return Pair.of(condExprVarUseTree, condExprVarUseNode);
     }