+++ framework/src/main/java/org/checkerframework/framework/flow/CFAbstractTransfer.java
@@ -667,7 +667,19 @@ public abstract class CFAbstractTransfer<
     @Override
     public TransferResult<V, S> visitTernaryExpression(
             TernaryExpressionNode n, TransferInput<V, S> p) {
-        return visitLocalVariable(n.getTernaryExpressionVar(), p);
+        TransferResult<V, S> result = super.visitTernaryExpression(n, p);
+        S thenStore = result.getThenStore();
+        S elseStore = result.getElseStore();
+
+        V thenValue = p.getValueOfSubNode(n.getThenOperand());
+        V elseValue = p.getValueOfSubNode(n.getElseOperand());
+        V resultValue = null;
+        if (thenValue != null && elseValue != null) {
+            // The resulting abstract value is the merge of the 'then' and 'else' branch.
+            resultValue = thenValue.leastUpperBound(elseValue);
+        }
+        V finishedValue = finishValue(resultValue, thenStore, elseStore);
+        return new ConditionalTransferResult<>(finishedValue, thenStore, elseStore);
     }
 
     @Override
@@ -817,7 +829,6 @@ public abstract class CFAbstractTransfer<
         Node lhs = n.getTarget();
         Node rhs = n.getExpression();
 
-        S store = in.getRegularStore();
         V rhsValue = in.getValueOfSubNode(rhs);
 
         /* NO-AFU
@@ -842,9 +853,20 @@ public abstract class CFAbstractTransfer<
                }
         */
 
-        processCommonAssignment(in, lhs, rhs, store, rhsValue);
-
-        return new RegularTransferResult<>(finishValue(rhsValue, store), store);
+        if (n.isSynthetic() && in.containsTwoStores()) {
+            // This is a synthetic assignment node created for a ternary expression. In this case
+            // the `then` and `else` store are not merged.
+            S thenStore = in.getThenStore();
+            S elseStore = in.getElseStore();
+            processCommonAssignment(in, lhs, rhs, thenStore, rhsValue);
+            processCommonAssignment(in, lhs, rhs, elseStore, rhsValue);
+            return new ConditionalTransferResult<>(
+                    finishValue(rhsValue, thenStore, elseStore), thenStore, elseStore);
+        } else {
+            S store = in.getRegularStore();
+            processCommonAssignment(in, lhs, rhs, store, rhsValue);
+            return new RegularTransferResult<>(finishValue(rhsValue, store), store);
+        }
     }
 
     @Override