+++ framework/src/main/java/org/checkerframework/framework/util/defaults/QualifierDefaults.java
@@ -489,12 +489,7 @@ public class QualifierDefaults {
     private Element nearestEnclosingExceptLocal(Tree tree) {
         TreePath path = atypeFactory.getPath(tree);
         if (path == null) {
-            Element element = atypeFactory.getEnclosingElementForArtificialTree(tree);
-            if (element != null) {
-                return element;
-            } else {
-                return TreeUtils.elementFromTree(tree);
-            }
+            return TreeUtils.elementFromTree(tree);
         }
 
         Tree prev = null;